% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grillade.R
\name{col}
\alias{col}
\title{Grillade column}
\usage{
col(..., col_width = NULL, row_height = NULL)
}
\arguments{
\item{...}{Content of the column, named arguments will be used as tag attributes.}

\item{col_width}{Column width.}

\item{row_height}{Row height.}
}
\value{
An element that can be used inside \code{\link{grillade}}.
}
\description{
Grillade column
}
