% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{grillade-dependency}
\alias{grillade-dependency}
\alias{html_dependency_grillade}
\alias{use_grillade}
\title{Grillade HTML dependency}
\usage{
html_dependency_grillade()

use_grillade()
}
\value{
an \code{\link[htmltools]{htmlDependency}}.
}
\description{
Those functions can be used to load
 \href{https://www.knacss.com/doc.html#grids}{grillade} explicitly.
 \code{use_grillade} can be used in a RMakrdown chunk.
}
