% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grillade-package.R
\name{grillade-exports}
\alias{grillade-exports}
\alias{tags}
\alias{tagList}
\title{grillade exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the grillade package.
}
