% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grillade.R
\name{grilladeW}
\alias{grilladeW}
\title{<Add Title>}
\usage{
grilladeW(
  ...,
  n_col = NULL,
  cols_width = NULL,
  .list = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
