% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{html_dependency_grillade}
\alias{html_dependency_grillade}
\title{Grillade HTML dependency}
\usage{
html_dependency_grillade()
}
\value{
an \code{\link[htmltools]{htmlDependency}}.
}
\description{
Grillade HTML dependency
}
