% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print-methods}
\alias{print-methods}
\alias{print.grillade}
\alias{print.knack}
\title{Print method for grillade/knack objects}
\usage{
\method{print}{grillade}(x, browsable = TRUE, ...)

\method{print}{knack}(x, ...)
}
\arguments{
\item{x}{A \code{\link{grillade}} or \code{\link{knack}} object.}

\item{browsable}{Make the result browsable.}

\item{...}{Additional arguments.}
}
\description{
Print method for grillade/knack objects
}
