% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.grillade}
\alias{print.grillade}
\title{Print method for grillade object}
\usage{
\method{print}{grillade}(x, ...)
}
\arguments{
\item{x}{A \code{\link{grillade}} object.}

\item{...}{Additional arguments.}
}
\description{
Print method for grillade object
}
