% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grillade.R
\name{col}
\alias{col}
\title{Grillade column}
\usage{
col(..., col_width = NULL, row_height = NULL)
}
\arguments{
\item{...}{Content of the column, named arguments will be used as tag attributes.}

\item{col_width}{Column width.}

\item{row_height}{Row height.}
}
\value{
An element that can be used inside \code{\link{grillade}}.
}
\description{
Grillade column
}
\examples{
library(grillade)

# For examples
box_example <- function(...) {
  tags$div(
    style = "box-sizing:border-box;border:5px solid #1C6EA4;height:100\%;",
    ...
  )
}

# Number of row taken by grid element
grillade(
  box_example("Column 1"),
  col(
    row_height = 2,
    box_example("Column 2")
  ),
  box_example("Column 3"),
  n_col = 2,
  gutter = TRUE
)


# Specify columns width
grillade(
  col(
    col_width = 3,
    box_example("Column 1")
  ),
  col(
    col_width = 2,
    box_example("Column 2")
  ),
  col(
    col_width = 2,
    box_example("Column 3")
  ),
  col(
    col_width = 3,
    box_example("Column 4")
  ),
  n_col = 5,
  gutter = TRUE
)


}
