% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{grillade-shiny}
\alias{grillade-shiny}
\alias{grilladeOutput}
\alias{renderGrillade}
\title{Grillade Output in Shiny}
\usage{
grilladeOutput(outputId, width = "100\%", ...)

renderGrillade(
  expr,
  n_col = NULL,
  cols_width = NULL,
  gutter = FALSE,
  env = parent.frame(),
  quoted = FALSE,
  outputArgs = list()
)
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width}{If not \code{NULL}, must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{...}{Other arguments to pass to the container tag function.
This is useful for providing additional classes for the tag.}

\item{expr}{An expression that generates a \code{\link{grillade}}.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\value{
An HTML output element that can be included in Shiny UI.
}
\description{
Grillade Output in Shiny
}
