% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{grillade}
\alias{grillade}
\title{Create a grillade (grid) of elements}
\usage{
grillade(..., n_col = NULL, cols_width = NULL, gutter = FALSE, .list = NULL)
}
\arguments{
\item{.list}{}
}
\value{

}
\description{
Display plots, htmlwidgets or other HTML components in a grid.
}
